refactor: decompose _layout, FloatingTabBar, ProfileForm, plates (FTA batch 6)
Extract sub-components and hooks to reduce file complexity:

- app/_layout.tsx (415→168L): Extract useAppInit hook and SCREEN_CONFIGS data
- FloatingTabBar.tsx (384→168L): Extract CenterButton and TabButton
- ProfileForm.tsx (364→193L): Extract useProfileForm hook and ActivityDropdown
- plates.tsx (367→257L): Extract usePlateCalculator hook, BarbellDiagram, StatusMessage

New files:
- hooks/useAppInit.ts, app/screen-config.ts
- components/floating-tab-bar/CenterButton.tsx, TabButton.tsx
- hooks/useProfileForm.ts, components/profile/ActivityDropdown.tsx
- hooks/usePlateCalculator.ts, components/plates/BarbellDiagram.tsx

Update source-reading tests to include extracted files.
Add 24 structural tests for batch 6 decomposition.
All 110 suites / 1737 tests pass.
